Oopsies.
So anyway, GenericLoader will still contain the finalized record type interface which inherits the IBasReader interface because all files being worked with in this program have at least a dbase file but most often contain a shapefile as well and the polished finalized fields are to be found in the recordloader type.

So far as the sql server tables I'll write a generic BO and CRUD generator as thats the easy part or use one of M$'s incredibly crappy failures.