started adding code for the descriptor file comparison to make sure the right goddamn data is in their database... still kind of taking their word though.
might add a shasum of the shapefile and dbf as well.
any duplicates means something is wrong.
with the shapefile this is more likely to work as that is going to be packed binary data... which is how i have to read the damn crap in. so there wont be things like spaces and the like to throw off a checksum.