the question is did zim supply a lie. see I could see a scatterplot style sightings of species gathering and creating boundaries of range arbitrarily to this, which would of course produce incomplete data but still usable.
so albers projections are where the species data is screwed up.
the distance between one point in latitude/longitude and another is slightly different just like gravity differs where you stand on the globe, only because the world is not a sphere but an ellipsoid. so.. not sure which direction to go really.

if i have to do any distance calculations the other system might be more useful. still useless without an elevation map though.