Still wondering if it would be useful to store coordinates in albers projections to make distance calculations simple. Its a thought. There has to be a reason cartographers in the usgs would prefer this.
Its funny to think how these boobs decided to add secondary meanings to more things than they themselves had any comprehension of but can recite things like this in bulk.

Personally never delved far enough into this except while working on this project to really look that closely at this concept.

But given the earth is an ellipsoid it would kind of make sense that a perfect sphere idea was crap.