jesus fing christ.
anyway, i do seem to think this worked.
tile sizes of 512x512 allow for enough variation on a color limited raster to allow for some really good compression.

and these people are still garbage and the year is not 2021.
they should all go live in their retarded version of the world alone and leave me out of this shit.